requirements.txt
accelerate==1.11.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.4
filelock==3.20.0
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
httpx-ws==0.8.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
lmstudio==1.5.0
markupsafe==3.0.3
mpmath==1.3.0
msgspec==0.19.0
networkx==3.5
numpy==2.3.4
openai>=1.0.0
packaging==25.0
protobuf==6.33.0
psutil==7.1.3
pymetasploit3>=1.0.3
python-dotenv>=1.0.0
python-nmap>=0.7.1
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.6.2
sentencepiece==0.2.1
setuptools==80.9.0
sniffio==1.3.1
sympy==1.14.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typing-extensions==4.15.0
urllib3==2.5.0
wsproto==1.2.0
